--- a/test/dense.jl
+++ b/test/dense.jl
@@ -193,4 +193,20 @@ end
   end
 end
 
+@testset "change backends" begin
+    a, b, c = [randn(5,5) for i=1:3]
+    backend_auto()
+    @test NDTensors.gemm_backend[] == :Auto
+    @test NDTensors.auto_select_backend(typeof.((a, b, c))...) == Val(:BLAS)
+    res1 = NDTensors._gemm!('N', 'N', 2.0, a, b, 0.2, copy(c))
+    backend_blas()
+    @test NDTensors.gemm_backend[] == :BLAS
+    res2 = NDTensors._gemm!('N', 'N', 2.0, a, b, 0.2, copy(c))
+    backend_generic()
+    @test NDTensors.gemm_backend[] == :Generic
+    res3 = NDTensors._gemm!('N', 'N', 2.0, a, b, 0.2, copy(c))
+    @test res1 == res2
+    @test res1 ≈ res3
+end
+
 nothing