For supporting different backends

```julia
using NDTensors, BenchmarkTools, TropicalNumbers
using ITensors

i = Index(1000)
j = Index(1000)
k = Index(1000)
t1 = ITensors.ITensor(TropicalF64, i, j)
t2 = ITensors.ITensor(TropicalF64, j, k)

using TropicalGEMM, Octavian
using LinearAlgebra: BlasFloat

@inline function NDTensors.auto_select_backend(::Type{<:StridedVecOrMat{<:Tropical{<:BlasFloat}}}, ::Type{<:StridedVecOrMat{<:Tropical{<:BlasFloat}}}, ::Type{<:StridedVecOrMat{<:Tropical{BlasFloat}}})
    Val(:Octavian)
end

function NDTensors._gemm!(::Val{:Octavian}, tA, tB, alpha,
                A::AbstractVecOrMat,
                B::AbstractVecOrMat,
                beta, C::AbstractVecOrMat)
    if tA == 'T'
        A = transpose(A)
    end
    if tB == 'T'
        B = transpose(B)
    end
    Octavian.matmul!(C, A, B, alpha, beta, C)
end

@benchmark t1 * t2
```
## A note on the benchmark results
Before using TropicalGEMM,
```julia

julia> @benchmark t1 * t2
BenchmarkTools.Trial: 
  memory estimate:  7.64 MiB
  allocs estimate:  66
  --------------
  minimum time:     2.287 s (0.00% GC)
  median time:      2.288 s (0.00% GC)
  mean time:        2.288 s (0.00% GC)
  maximum time:     2.288 s (0.00% GC)
  --------------
  samples:          3
  evals/sample:     1
```

After using TropicalGEMM,

```julia
julia> @benchmark t1 * t2
BenchmarkTools.Trial: 
  memory estimate:  7.63 MiB
  allocs estimate:  60
  --------------
  minimum time:     67.562 ms (0.00% GC)
  median time:      67.964 ms (0.00% GC)
  mean time:        68.596 ms (0.25% GC)
  maximum time:     72.188 ms (1.34% GC)
  --------------
  samples:          73
  evals/sample:     1
```

## A note on the overheads 
```julia
julia> backend_generic()
:Generic

julia> @benchmark NDTensors._gemm!('N', 'N', 1.0, $a, $b, 0.0, $c)
BenchmarkTools.Trial: 
  memory estimate:  96 bytes
  allocs estimate:  1
  --------------
  minimum time:     302.398 ns (0.00% GC)
  median time:      306.681 ns (0.00% GC)
  mean time:        314.784 ns (0.92% GC)
  maximum time:     6.286 μs (94.60% GC)
  --------------
  samples:          10000
  evals/sample:     251

julia> backend_auto()
:Auto

julia> @benchmark NDTensors._gemm!('N', 'N', 1.0, $a, $b, 0.0, $c)
BenchmarkTools.Trial: 
  memory estimate:  0 bytes
  allocs estimate:  0
  --------------
  minimum time:     112.454 ns (0.00% GC)
  median time:      113.455 ns (0.00% GC)
  mean time:        114.501 ns (0.00% GC)
  maximum time:     187.419 ns (0.00% GC)
  --------------
  samples:          10000
  evals/sample:     923
```