@GiggleLiu 2021-04-02
> It would be nice to have a similar interface at the level of the contract! function for switching between contraction backends

Agree, let's do it later.

> overload in a file that only gets included through Requires when Octavian is loaded?

Sounds good. Added.

> Is that overhead related to the dynamic construction of Val in this line: Val(gemm_backend[])? I guess that is unavoidable if you want the backend to be externally extendable, but at least through the "automatic" backend selection that overhead isn't there as long as auto_select_backend is written in a type stable way.

Exactly.

> specialized Val-like type called MatMulBackend

Sounds good, I used GemmBackend to make the naming more consistent.

> Finally, we probably want a simpler interface for the _gemm! function to make this overloading nicers. There shouldn't be a need for people to have the lines:

We can not avoid these lines if we call it gemm. To make it less terrifying, I made it look shorter.

> Finally, a small interface thing, it would be better to rename NDTensors.auto_select_backend to something like NDTensors.auto_select_matmul_backend to allow for a future function NDTensors.auto_select_contract_backend. I also think it would be better to have a more general function like set_matmul_backend(s::String) instead of the hard-coded functions backend_auto(). I can never decide if functions like that should be set_matmul_backend(s) or set_matmul_backend!(s)... We have also generally preferred using Strings instead of Symbols whenever possible in the external interface since they are more familiar to a wider audience, and I don't think there is any advantage to using a Symbol in this case, at least in the interface set_matmul_backend("auto") or set_matmul_backend("blas"). We can use Symbols internally for efficiency.

We can do it whenever we need it. A deprecation warning will help people migrate.

A general comment
--------------------------------
I feel `GEMM` is a lower level interface, while `matmul!` and `contract!` are higher level interfaces. This is why I think `GEMM` is better name for backends. It is wield to specify both `matmul` and `contract!` methods, because `contract!` includes `matmul!`. Anyway, feel free to change.