@mtfishman 2021-04-02
> I feel `GEMM` is a lower level interface, while `matmul!` and `contract!` are higher level interfaces. This is why I think `GEMM` is better name for backends. It is wield to specify both `matmul` and `contract!` methods, because `contract!` includes `matmul!`. Anyway, feel free to change.

My point with preferring `matmul!` over `gemm!` is that they are both just as general interfaces for the same operations, but the Julia style of `matmul!`/`mul!` is much cleaner and simpler. So a goal would be to just use `matmul!`/`mul!` everywhere anyway. I don't think there was a point of calling `gemm!` directly in `contract!`, that was just historical and we could just as easily call Julia's 5-arg `mul!` as you changed in the generic case in this PR. 

Regarding overloading `contract!`, there are certain contraction backends like [TBLIS](https://github.com/devinamatthews/tblis) or [GETT](https://github.com/HPAC/tccg) that never call a gemm, and instead perform tensor contractions directly without any intermediate permutations. In that case, you would have to overload `contract!` directly since it wouldn't even reach use a gemm call. Even besides that point, `NDTensors.auto_select_backend` is just a very general name and it wouldn't be clear what operation's backend it was referring to (for example, we also have a variety of SVD backends, it isn't clear that this isn't controlling that as well).

Anyway, we can merge this as is and I'm happy to clean it up, but I prefer to get the interface as good as possible the first time to avoid the annoyance of deprecations, when possible.